+++ src/components/BootstrapBlazor.ImageCropper/BootstrapBlazor.ImageCropper.csproj
@@ -1,41 +1,16 @@
 <Project Sdk="Microsoft.NET.Sdk.Razor">
-    <PropertyGroup>
-        <TargetFrameworks>net6.0;net7.0;net8.0</TargetFrameworks>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-    </PropertyGroup>
 
-    <PropertyGroup>
-        <Version>0.0.4</Version>
-        <PackageProjectUrl>https://github.com/densen2014/BootstrapBlazor.ImageCropper</PackageProjectUrl>
-        <RepositoryUrl>https://github.com/densen2014/BootstrapBlazor.ImageCropper</RepositoryUrl>
-        <RepositoryType>git</RepositoryType>
-        <PackageTags>Bootstrap Blazor ImageCropper 图像裁剪 组件  </PackageTags>
-        <Description>Bootstrap UI ImageCropper components experience</Description>
-        <AssemblyName>BootstrapBlazor.ImageCropper</AssemblyName>
-        <RootNamespace>BootstrapBlazor.ImageCropper</RootNamespace>
-        <AssemblyOriginatorKeyFile>..\keys\Longbow.Utility.snk</AssemblyOriginatorKeyFile>
-        <SignAssembly>True</SignAssembly>
-        <OutputPath>..\..\..\Bin</OutputPath>
-    </PropertyGroup>
+  <PropertyGroup>
+    <Version>8.1.0</Version>
+  </PropertyGroup>
 
-    <ItemGroup Condition="'$(TargetFramework)' == 'net6.0'">
-        <PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="6.0.10" />
-    </ItemGroup>
+  <PropertyGroup>
+    <PackageTags>Bootstrap Blazor WebAssembly wasm UI Components ImageCropper</PackageTags>
+    <Description>Bootstrap UI components extensions of ImageCropper</Description>
+  </PropertyGroup>
 
-    <ItemGroup Condition="'$(TargetFramework)' == 'net7.0'">
-        <PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="7.0" />
-    </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="BootstrapBlazor" Version="8.10.2" />
+  </ItemGroup>
 
-    <ItemGroup Condition="'$(TargetFramework)' == 'net8.0'">
-        <PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="8.0" />
-    </ItemGroup>
-    
-    <!-- 通过移除正常的引用并替换成固定版本的引用，达到无论如何更新都无法生效的目的。 -->
-    <Target Condition="'$(TargetFramework)' == 'net6.0'" Name="ReferenceStaticLegacyPackage" BeforeTargets="CollectPackageReferences">
-        <ItemGroup>
-            <PackageReference Remove="Microsoft.AspNetCore.Components.Web" />
-            <PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="6.0" />
-        </ItemGroup>
-    </Target>
 </Project>
