+++ src/components/BootstrapBlazor.ImageCropper/ImageCropper.razor
@@ -1,28 +1,7 @@
 ﻿@namespace BootstrapBlazor.Components
+@inherits BootstrapModuleComponentBase
+@attribute [JSModuleAutoLoader("./_content/BootstrapBlazor.ImageCropper/ImageCropper.razor.js", JSObjectReference = true)]
 
-<link href="@CssPath" rel="stylesheet">
-
-<div @ref="Element">
-
-    <div class="containerTop" data-crop-shape="@CropShape.ToString().ToLower()">
-        <div>
-            <img class="@CssClass" data-action="image" src="@Url" alt="Picture">
-        </div>
-        @if (DefaultButton)
-        {
-            <div class="input-group">
-                <button type="button" onclick="@(() => Crop())" class="btn btn-primary">@ConfirmBtnTitle</button>
-                <button type="button" onclick="@(() => Reset())" class="btn btn-primary">@ResetBtnTitle</button>
-            </div>
-        }
-        @if (Preview)
-        {
-            <h3>@PreviewTitle</h3>
-            <div data-action="result"></div>
-        }
-
-
-    </div>
-
+<div @attributes="@AdditionalAttributes" class="@ClassString" id="@Id">
+    <img src="@Url" class="bb-cropper-image" />
 </div>
-
