+++ src/components/BootstrapBlazor.ImageCropper/ImageCropper.razor.cs
@@ -1,192 +1,67 @@
-﻿// ********************************** 
-// Densen Informatica 中讯科技 
-// 作者：Alex Chow
-// e-mail:zhouchuanglin@gmail.com 
-// **********************************
+﻿// Copyright (c) Argo Zhang (argo@163.com). All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+// Website: https://www.blazor.zone or https://argozhang.github.io/
 
 using Microsoft.AspNetCore.Components;
-using Microsoft.JSInterop;
-using System.Diagnostics.CodeAnalysis;
-using System.Text.RegularExpressions;
 
 namespace BootstrapBlazor.Components;
 
 /// <summary>
 /// 视频播放器 ImageCropper 组件
 /// </summary>
-public partial class ImageCropper : IAsyncDisposable
+public partial class ImageCropper
 {
-    [NotNull]
-    private IJSObjectReference? Module { get; set; }
-
-    [Inject, NotNull]
-    protected IJSRuntime? JSRuntime { get; set; }
-
-    private DotNetObjectReference<ImageCropper>? Instance { get; set; }
-
     /// <summary>
-    /// UI界面元素的引用对象
-    /// </summary>
-    private ElementReference Element { get; set; }
-
-    /// <summary>
-    /// 获得/设置 用于渲染的文件流,为空则用 FileName 参数读取文件
-    /// </summary>
-    [Parameter]
-    public Stream? Stream { get; set; }
-
-    private byte[]? streamCache { get; set; }
-
-    /// <summary>
-    /// 获得/设置 图片URL/Base64 dataurl
+    /// 获得/设置 图片地址 URL
     /// </summary>
     [Parameter]
     public string? Url { get; set; }
 
     /// <summary>
-    /// 确认按钮文本/Confirm button title
+    /// 获得/设置 剪裁结果回调方法
     /// </summary>
     [Parameter]
-    public string ConfirmBtnTitle { get; set; } = "OK";
-
-    /// <summary>
-    /// 复位按钮文本/Reset button title
-    /// </summary>
-    [Parameter]
-    public string ResetBtnTitle { get; set; } = "复位";
-
-    /// <summary>
-    /// 预览文本/Preview title
-    /// </summary>
-    [Parameter]
-    public string PreviewTitle { get; set; } = "预览";
-
-    /// <summary>
-    /// 显示默认按钮/Show default button
-    /// </summary>
-    [Parameter]
-    public bool DefaultButton { get; set; } = true;
-
-    /// <summary>
-    /// 显示剪裁后预览/Preview image afrer cropper 
-    /// </summary>
-    [Parameter]
-    public bool Preview { get; set; } = true;
-
-    /// <summary>
-    /// 获得/设置 读取本地文件路径
-    /// </summary> 
-    [Parameter]
-    public string? LocalFileName { get; set; }
-
-    /// <summary>
-    /// 自定义图片 CSS
-    /// </summary>
-    [Parameter]
-    public string? CssClass { get; set; }
-
-    /// <summary>
-    /// 自定义CSS,,默认为内置 cropper.css
-    /// </summary>
-    [Parameter]
-    public string? CssPath { get; set; } = "./_content/BootstrapBlazor.ImageCropper/cropper.css" + "?v=" + Ver;
-
-    /// <summary>
-    /// 自定义cropper.js路径,默认为null,使用内置 cropper.js
-    /// </summary>
-    [Parameter]
-    public string? ModulePath { get; set; }
-
-    /// <summary>
-    /// 获得/设置 剪裁结果回调方法/Cropper result callback method
-    /// </summary>
-    [Parameter]
-    public Func<Stream, Task>? OnResult { get; set; }
-
-    /// <summary>
-    /// 获得/设置 剪裁结果base64回调方法
-    /// </summary>
-    [Parameter]
-    public Func<string, Task>? OnBase64Result { get; set; }
-
-    /// <summary>
-    /// 获得/设置 错误回调方法
-    /// </summary>
-    [Parameter]
-    public Func<string, Task>? OnError { get; set; }
-
-    private static string? Ver { get; set; } = System.Reflection.Assembly.GetExecutingAssembly().GetName().Version?.ToString();
-
-    [NotNull]
-    private string? Id { get; set; }
-    private string? ErrorMessage { get; set; }
+    public Func<ImageCropperResult, Task>? OnCropAsync { get; set; }
 
     /// <summary>
     /// 获取/设置 裁剪选项
     /// </summary>
     [Parameter]
-    public CropperOption Options { get; set; } = new();
+    public ImageCropperOption? Options { get; set; }
 
     /// <summary>
-    /// 获取/设置 裁剪形状（矩形/圆形）
+    /// 获取/设置 裁剪形状（矩形/圆形）默认 <see cref="ImageCropperShape.Rectangle"/>
     /// </summary>
     [Parameter]
-    public CropShape CropShape { get; set; } = CropShape.Rectangle;
+    public ImageCropperShape CropperShape { get; set; }
+
+    private string? ClassString => CssBuilder.Default("bb-cropper")
+        .AddClass("is-round", CropperShape == ImageCropperShape.Round)
+        .AddClassFromAttributes(AdditionalAttributes)
+        .Build();
 
     /// <summary>
     /// <inheritdoc/>
     /// </summary>
-    protected override void OnInitialized()
-    {
-        base.OnInitialized();
-
-        Id = $"vp_{GetHashCode()}";
-    }
-
-    /// <summary>
-    /// <inheritdoc/>
-    /// </summary>
-    /// <param name="firstRender"></param>
     /// <returns></returns>
-    protected override async Task OnAfterRenderAsync(bool firstRender)
-    {
-        try
-        {
-            if (firstRender)
-            {
-                ModulePath = ModulePath ?? $"./_content/BootstrapBlazor.ImageCropper/cropper.js" + "?v=" + Ver;
-
-                await JSRuntime.InvokeAsync<IJSObjectReference>("import", ModulePath);
-
-                Module = await JSRuntime.InvokeAsync<IJSObjectReference>("import", "./_content/BootstrapBlazor.ImageCropper/ImageCropper.razor.js" + "?v=" + Ver);
-                Instance = DotNetObjectReference.Create(this);
-                await Module!.InvokeVoidAsync("init", Instance, Element, Options);
-                //await Module!.InvokeVoidAsync("changeAvatar", Instance, Element);
-
-            }
-        }
-        catch (Exception e)
-        {
-            OnError?.Invoke(e.Message);
-        }
-    }
+    protected override Task InvokeInitAsync() => InvokeVoidAsync("init", Id, Interop, Options);
 
     /// <summary>
-    /// 剪裁,返回 base64, 并执行 OnResult/OnBase64Result 回调
+    /// 剪裁方法 返回 base64 字符串
     /// </summary>
     /// <returns>base64</returns>
-    public async Task<string> Crop()
+    public async Task<string?> Crop()
     {
-        var base64encodedstring = await Module!.InvokeAsync<string>("crop");
-        await GetResult(base64encodedstring);
-        return base64encodedstring;
+        var result = await InvokeAsync<string?>(Id, "crop");
+        if (!string.IsNullOrEmpty(result))
+        {
+            if (OnCropAsync != null)
+            {
+                await OnCropAsync(new ImageCropperResult(result));
+            }
+        }
+        return result;
     }
-
-    /// <summary>
-    /// 剪裁,返回 Stream
-    /// </summary>
-    /// <returns>Stream</returns>
-    public async Task<Stream> CropToStream() => DataUrl2Stream(await Module!.InvokeAsync<string>("crop"));
 
     /// <summary>
     /// 替换图片
@@ -240,159 +115,9 @@ public partial class ImageCropper : IAsyncDisposable
     public async Task Destroy() => await Module!.InvokeVoidAsync("destroy");
 
     /// <summary>
-    /// 旋转, 90, 180, 270, -90 ...
+    /// 旋转图片方法
     /// </summary>
-    /// <param name="degree"></param>
+    /// <param name="angle">旋转角度</param>
     /// <returns></returns>
-    public async Task Rotate(int degree) => await Module!.InvokeVoidAsync("rotate", degree);
-
-    //[JSInvokable]
-    public async Task GetResult(string base64encodedstring)
-    {
-        try
-        {
-            if (OnResult != null)
-            {
-                await OnResult.Invoke(DataUrl2Stream(base64encodedstring));
-            }
-            if (OnBase64Result != null)
-            {
-                await OnBase64Result.Invoke(base64encodedstring);
-            }
-        }
-        catch (Exception e)
-        {
-            if (OnError != null)
-            {
-                await OnError.Invoke(e.Message);
-            }
-        }
-    }
-
-    /// <summary>
-    /// 选择图片解码 / Select picture decoding
-    /// </summary>
-    /// <param name="dataUrl">可选直接解码 Base64, DataUrl 格式</param>
-    /// <returns></returns>
-    public async Task DecodeFromImage(string? dataUrl = null)
-    {
-        if (dataUrl != null && !dataUrl.StartsWith("data:image"))
-        {
-            dataUrl = "data:image/jpeg;base64," + dataUrl;
-        }
-        await Module!.InvokeVoidAsync("DecodeFormImage", dataUrl);
-    }
-
-    /// <summary>
-    /// 从 DataUrl 转换为 Stream
-    /// <para>Convert from a DataUrl to an Stream</para>
-    /// </summary>
-    /// <param name="base64encodedstring"></param>
-    /// <returns></returns>
-    public static Stream DataUrl2Stream(string base64encodedstring)
-    {
-        var base64Data = Regex.Match(base64encodedstring, @"data:image/(?<type>.+?),(?<data>.+)").Groups["data"].Value;
-        var bytes = Convert.FromBase64String(base64Data);
-        var stream = new MemoryStream(bytes);
-        return stream;
-    }
-
-    public virtual async Task Refresh()
-    {
-        ErrorMessage = null;
-        try
-        {
-            if (Stream != null)
-            {
-                await ShowPdf(Stream);
-            }
-            else if (!string.IsNullOrEmpty(LocalFileName) && File.Exists(LocalFileName))
-            {
-                var streamLocal = new FileStream(LocalFileName, FileMode.Open, FileAccess.Read);
-                if (streamLocal != null)
-                {
-                    await ShowPdf(streamLocal);
-                }
-                else
-                {
-                    ErrorMessage = "No data";
-                }
-            }
-            else if (!string.IsNullOrEmpty(Url) && Url.StartsWith("http"))
-            {
-                var client = new HttpClient();
-                var stream = await client.GetStreamAsync(Url);
-                if (stream != null)
-                {
-                    await ShowPdf(stream);
-                }
-                else
-                {
-                    ErrorMessage = "No data";
-                }
-            }
-
-        }
-        catch (Exception e)
-        {
-            ErrorMessage = e.Message;
-        }
-        StateHasChanged();
-
-    }
-
-
-    /// <summary>
-    /// 打开 stream
-    /// </summary>
-    public virtual async Task ShowPdf(Stream stream)
-    {
-        if (Module == null)
-        {
-            Stream = stream;
-        }
-        else
-        {
-            if (streamCache == null)
-            {
-                streamCache = new byte[stream.Length];
-                stream.Read(streamCache, 0, (int)stream.Length);
-            }
-            if (streamCache != null)
-            {
-                Url = null;
-                using var streamRef = new DotNetStreamReference(new MemoryStream(streamCache));
-                await Module!.InvokeVoidAsync("showPdf", Url, Element, streamRef);
-            }
-        }
-    }
-
-    /// <summary>
-    /// 错误回调方法
-    /// </summary>
-    /// <param name="error"></param>
-    /// <returns></returns>
-    [JSInvokable]
-    public async Task GetError(string error)
-    {
-        if (OnError != null)
-        {
-            await OnError.Invoke(error);
-        }
-    }
-
-    /// <summary>
-    /// <inheritdoc/>
-    /// </summary>
-    /// <returns></returns>
-    public async ValueTask DisposeAsync()
-    {
-        if (Module is not null)
-        {
-            await Module.InvokeVoidAsync("destroyMod");
-            await Module.DisposeAsync();
-        }
-        GC.SuppressFinalize(this);
-    }
-
+    public async Task Rotate(int angle) => await InvokeVoidAsync(Id, "rotate", angle);
 }
