+++ src/components/BootstrapBlazor.ImageCropper/wwwroot/cropper.bundle.css
@@ -1,0 +1,16 @@
+﻿@import url('./cropper.min.css');
+
+.bb-cropper {
+    max-width: 640px;
+    width: 100%;
+    height: 100%;
+}
+
+    .bb-cropper .bb-cropper-image {
+        width: 100%;
+    }
+
+    .bb-cropper.is-round .cropper-view-box,
+    .bb-cropper.is-round .cropper-face {
+        border-radius: 50%;
+    }
